+++ app/src/main/java/br/com/kosawalabs/apprecommendation/presentation/list/contract/AppListModel.java
@@ -9,7 +9,7 @@ public interface AppListModel {
 
     void setRecommended(boolean recommended);
 
-    long getPageSize();
+    int getPageSize();
 
     boolean hasStopLoading();
 }
